Calculates posterior predictive checks (Bayesian p-values).

This function is in development. Use at your own risk.

sma_assess(
  object,
  expr,
  data,
  code,
  constants = nlist(),
  nsamples = 100L,
  statistic = "FT"
)

Arguments

object
An mcmcr object of results

expr
An expression to calculate the expectations, e.g. "X = mu" if the expectation of X is mu

data
A dataset

code
Code to simulate data

constants
Constants to pass with the code

nsamples
Number of samples to take from the posterior distribution

statistic
A string, one of "FT" (Freeman-Tukey), "LR" (Likelihood Ratio), or "chi2" (Chi-square)

Value

A flag.

Examples

code="for(i in 1:10){
a[i] ~ dpois(mu)}"
sims <- sims::sims_simulate(code, parameters=nlist(mu=5), nsims=1)
res <- sma_analyse(sims, 
                           code, 
                           "mu ~ dunif(0,10)", 
                           monitor = "mu",
                           deviance=FALSE,
                           mode=sma_set_mode("quick", n.save=300, max.iter=300))

#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 10
#>    Unobserved stochastic nodes: 1
#>    Total graph size: 13
#> 
#> Initializing model
#> 

sma_assess(res, expr="a = rep(mu, 10)", sims, code, nsamples=10)

#> Warning: This function is in development and has not been tested. Use at your own risks.

#> $D.data
#> $a
#>  [1] 0.01073261 0.16350307 0.26088313 0.16350307 0.01073261 0.52105729
#>  [7] 0.09916034 0.16350307 0.26088313 0.01073261
#> 
#> nchains:  1 
#> niters:  10 
#> 
#> 
#> $D.sims
#> $a
#>  [1] 0.08956132 0.22635024 0.21180302 0.04172012 0.10201772 0.10956136
#>  [7] 0.04064688 0.05036129 0.08516976 0.19882793
#> 
#> nchains:  1 
#> niters:  10 
#> 
#> 
#> $p.value
#> $p.value$a
#>  [1] 0.80 0.65 0.60 0.35 0.75 0.35 0.30 0.25 0.20 0.85
#> 
#> 
#> $fit
#> $fit$a
#> [1] 0.1813305
#> 
#>