Apply R code to derive new variables.

Apply R code to derive new variables.

sma_derive(
  object = NULL,
  code,
  monitor = ".*",
  values = list(),
  progress = FALSE,
  options = furrr::furrr_options()
)

Arguments

object

An object of class (or that can be coerced to) mcmcrs, mcmcr, nlists or nlist.

code

A string of R code to derive posterior samples for new parameters. E.g. "var = sigma^2".

monitor

A character vector (or regular expression if a string) specifying the names of the variables in object and/or code to monitor. By default all variables are included.

values

A named list of additional R objects to evaluate in the R expression.

progress

A flag specifying whether to print a progress bar.

options

The future specific options to use with the workers.

Value

An object of the same class as object

Examples

set.seed(10L)
code <- "for(i in 1:10){x[i] ~ dnorm(0,1/variance)}"
parameters = nlist(variance=4)
dat <- sims::sims_simulate(code, parameters = parameters, nsims=2)
res <- sma_analyse(dat, code, code.add = "variance ~ dunif(0,10)", 
mode=sma_set_mode("quick"), monitor="variance")
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 10
#>    Unobserved stochastic nodes: 1
#>    Total graph size: 15
#> 
#> Initializing model
#> 
#> 
#> Max r.hat= 1.219
#> Min ess= 9
#> Module dic unloaded
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 10
#>    Unobserved stochastic nodes: 1
#>    Total graph size: 15
#> 
#> Initializing model
#> 
#> 
#> Max r.hat= 1
#> Min ess= 15
#> Module dic unloaded
#> module dic loaded
#> module dic loaded
sma_derive(res, "sd=sqrt(variance)")
#> $mcmcr1
#> $deviance
#> [1] 36.87924
#> 
#> nchains:  2 
#> niters:  10 
#> 
#> $sd
#> [1] 1.81186
#> 
#> nchains:  2 
#> niters:  10 
#> 
#> $variance
#> [1] 3.283009
#> 
#> nchains:  2 
#> niters:  10 
#> 
#> 
#> $mcmcr2
#> $deviance
#> [1] 38.82901
#> 
#> nchains:  2 
#> niters:  10 
#> 
#> $sd
#> [1] 1.66555
#> 
#> nchains:  2 
#> niters:  10 
#> 
#> $variance
#> [1] 2.774295
#> 
#> nchains:  2 
#> niters:  10 
#> 
#> 
sma_derive(parameters, "sd=sqrt(variance)")
#> $sd
#> [1] 2
#> 
#> $variance
#> [1] 4
#> 
#> an nlist object with 2 numeric elements