Read files

Read output files

sma_read_files(fun, path = ".", folder = "analysis0000001", id = 1)

Arguments

fun
The function from which to read the ouput file. Options are sma_analyse_files, sma_derive_files and sma_evaluate_files.

path
A string. The object is read using this path.

folder
A string for the name of the folder within the path that contains the analysis.

id
An integer for the index of the object to extract.

Value

The object read from files.

Examples

params <- nlist(mu=0)
sims::sims_simulate("a ~ dnorm(mu, 1)", 
                    parameters = params, 
                    nsims = 5,
                    save = TRUE,
                    path=tempdir(),
                    exist = NA)
#> [1] TRUE
sma_analyse_files(code = "a ~ dnorm(mu, 1)
                          mu ~ dunif(-3,3)",
                  mode = sma_set_mode("quick"),
                  monitor = "mu",
                  path = tempdir())
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 1
#>    Unobserved stochastic nodes: 1
#>    Total graph size: 5
#> 
#> Initializing model
#> 
#> 
#> Max r.hat= 1
#> Min ess= 20
#> Module dic unloaded
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 1
#>    Unobserved stochastic nodes: 1
#>    Total graph size: 5
#> 
#> Initializing model
#> 
#> 
#> Max r.hat= 1
#> Min ess= 20
#> Module dic unloaded
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 1
#>    Unobserved stochastic nodes: 1
#>    Total graph size: 5
#> 
#> Initializing model
#> 
#> 
#> Max r.hat= 1
#> Min ess= 14
#> Module dic unloaded
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 1
#>    Unobserved stochastic nodes: 1
#>    Total graph size: 5
#> 
#> Initializing model
#> 
#> 
#> Max r.hat= 1.062
#> Min ess= 20
#> Module dic unloaded
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 1
#>    Unobserved stochastic nodes: 1
#>    Total graph size: 5
#> 
#> Initializing model
#> 
#> 
#> Max r.hat= 1.518
#> Min ess= 8
#> Module dic unloaded
#> module dic loaded
#> module dic loaded
#> module dic loaded
#> module dic loaded
#> module dic loaded
#> character(0)
#> ---
#> 
#> ✔ data0000001.rds [00:00:00.032]
#> ✔ data0000002.rds [00:00:00.038]
#> ✔ data0000003.rds [00:00:00.032]
#> ✔ data0000004.rds [00:00:00.036]
#> ✔ data0000005.rds [00:00:00.032]
#> Success: 5
#> Failure: 0
#> Remaining: 0
#> 
#> [1] FALSE
sma_evaluate_files(parameters = params, 
                   path = tempdir())
sma_read_files(sma_evaluate_files, path = tempdir())
#>   term       bias cpQuantile       mse
#> 1   mu -0.2110081          1 0.5820712