Set mode for data analysis

Set mode for data analysis. The algorithm starts with an adaptation phase of n.adapt iterations. Then n.chains are ran for n.save iterations without thinning. The algorithm stops if convergence was achieved according to r.hat and the minimum effective sampling rate was achieved or max.time was reached or is projected to be reached in the next round of iterations. If the algorithm is not stopped, the iterations so far are discarded as burnin. The number of iterations is doubled from the previous round and thinning is multiplied by 2. The stop criteria is applied again and so on.

sma_set_mode(
  mode = "report",
  n.chains,
  n.adapt,
  n.save,
  max.iter,
  r.hat,
  r.hat.nodes,
  ess,
  ess.nodes,
  max.time,
  units
)

Arguments

mode
A string of the analysis mode. One of "quick", "report" or "paper". The mode determines the values of the parameters below unless they are specified.

n.chains
A count for number of chains.

n.adapt
A count for the number of adaptations.

n.save
A count for the number of (potentially thinned) samples to save.

max.iter
A count for the maximum number of iterations in total per chain.

r.hat
A number specifying the rhat threshold.

r.hat.nodes
A character vector of nodes to be used to assess r.hat

ess
A number specifying the minimum effective sample size.

ess.nodes
A character vector of nodes to be used to assess ess.

max.time
A number specifying the maximum time to spend on analysis.

units
A character string specifying the units of time for max.time. See difftime. The possible modes are as follows:

'quick'

To rapidly identify problems with a model definition.

'report'

To produce results for a report.

'paper'

To produce results for a peer-reviewed paper.

Examples

if (FALSE) {
sma_set_mode("quick")
}